[!!!][TASK] Split SolrService into SolrAdminService, SolrReadService and SolrWriteService
This pr:

* Extracts the SolrAdminService, SolrReadService and SolrWriteService from the SolrService.
* Changes the SolrConnectionManager to return a SolrConnection that could be used to access these services.
* Allows as followup to differnciate between read and write connections an to send them to different solr servers (master/slave)

Impact:

The ConnectionManager now return a SolrConnection that has the methods "getReadService", "getWriteService" and get "getAdminService".
These Services contain the corresponding method from the previous SolrService.

Fixes: #1745
