+++ Tests/Integration/System/Solr/Service/Fixtures/testpdf.pdf
